+++ R/zzz.R
@@ -4,5 +4,7 @@
 
 .onLoad = function(libname = find.package("PACVr"), pkgname = "PACVr") {
     # nothing
+    library(dplyr)
+    library(logger)
     library(RCircos)
 }