+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 CHANGELOG
 ---------
 
-#### Version 1.0.4 (2023.11.21)
-* Fix error caused by function masking
+#### Version 1.0.4 (2023.11.23)
+* Implemented a logger
+* Several bug fixes (e.g., error caused by function masking)
 * Resolve warning caused by depreciated 'dplyr' use
 
 #### Version 1.0.3 (2023.10.05)